% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_name_fishbase.R
\name{check_name_fishbase}
\alias{check_name_fishbase}
\title{Returns error if name is incorrect}
\usage{
check_name_fishbase(sp)
}
\arguments{
\item{sp}{A character value containing the species name}

\item{sp}{A character value containing the species name}
}
\description{
This is a wrapper function to which will return an error
(via \code{\link[fishflux]{name_errors}}) if the provided species
name is wrong.
}
\examples{

check_name_fishbase("Lutjanus griseus")
plyr::ldply(lapply(c("Chlorurus spilurus","Zebrasoma copas"), check_name_fishbase))
Returns error if name is incorrect

This is a wrapper function to which will return an error (via
\\code{\\link[fishflux]{name_errors}}) if the provided species name is wrong.




check_name_fishbase("Lutjanus griseus")
plyr::ldply(lapply(c("Chlorurus spilurus","Zebrasoma copas"), check_name_fishbase))

}
\keyword{fish,}
\keyword{fishbase}
\keyword{taxonomy,}
