% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnp_fishf.R
\name{cnp_fishf}
\alias{cnp_fishf}
\title{A function to estimate CNP and sd for the highest phylogenetic level
available}
\usage{
cnp_fishf(sp = NA, family = NA)
}
\arguments{
\item{sp}{A charachter value containing the species name}

\item{family}{A charachter value containing family name}
}
\description{
A function to estimate CNP and sd for the highest phylogenetic level
available
}
\examples{


fishflux::cnp_fishf("Lutjanus griseus")


}
\keyword{CNP}
\keyword{fish,}
