% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnp_model.R
\name{cnp_model}
\alias{cnp_model}
\title{A function to predict N and P excretion, egestion, ingestion rate, etc}
\usage{
cnp_model(TL, AEn = 0.8, AEp = 0.7, AEc = 0.8, Fn, Fp, Fc, t0, Linf,
  k, f, asp, troph, lwa, lwb, w_prop = 0.27, temp = 27, Tn = 0.0057,
  Tp = 0.00016, N, C, P, a = 0.83, B0 = NA)
}
\arguments{
\item{TL}{total length of a fish in cm}

\item{AEn}{N-specific assimilation efficiency}

\item{AEp}{P-specific assimilation efficiency}

\item{AEc}{C-specific assimilation efficiency}

\item{Fn}{N content of food in percent}

\item{Fp}{P content of food in percent}

\item{Fc}{C content of food in percent}

\item{t0}{Von Bertalanffy growth parameter}

\item{Linf}{Von Bertalanffy growth parameter (Should be higher than largest
fish)}

\item{k}{Von Bertalanffy growth parameter}

\item{f}{activity scope. Ranges from 1 to 4}

\item{asp}{aspect ratio of caudal fin}

\item{troph}{trophic level}

\item{lwa, lwb}{parameters for length-weight relationship}

\item{w_prop}{ratio between dry weight and wet weight of fish}

\item{temp}{temperature in degrees celcius}

\item{Tn, Tp}{N,P specific biomass turnover rates}

\item{N, C, P}{N,C,P content of fish in percent}

\item{a}{Resting metabolic rate mass-scaling exponent}

\item{B0}{Constant for resting metabolic rate. If NA, function will
calculate an average.}
}
\description{
This function combines MTE and stoichiometric theory in order to predict
necessary ingestion and excretion processes.
}
\details{
Returns a list, containing a dataframe with with all model components
(result), specifications of metabolism, and the limiting element.
}
\examples{


sp <- "Scarus psittacus"
para <- model_parameters(sp,"Scaridae")$parameters
para
test <- fishflux::cnp_model(sp = sp,TL = 10:19,Fn = 1.3,Fp = 0.1,Fc = 65,
               t0 = para$t0,Linf = para$Linf,k = para$k,Linf = para$Linf,
               asp = para$asp,troph = para$troph,f = 3,w_prop = para$w_prop,
               lwa = para$lwa,lwb = para$lwb,temp = 27,N = para$N,P=para$P,C=para$C)

}
\keyword{excretion}
\keyword{fish,}
\keyword{stoichiometry,}
