% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnp_model_mcmc.R
\name{cnp_model_mcmc}
\alias{cnp_model_mcmc}
\title{A function to predict N and P excretion, CNP egestion, CNP ingestion rate, using MCMC and stan}
\usage{
cnp_model_mcmc(TL, param, iter = 1000, cor = list(ro_Qc_Qn = 0.5,
  ro_Qc_Qp = -0.3, ro_Qn_Qp = -0.2, ro_Fc_Fn = 0.2, ro_Fc_Fp = -0.1,
  ro_Fn_Fp = -0.1, ro_lwa_lwb = 0.9, ro_a_B0 = 0.9), ...)
}
\arguments{
\item{TL}{total length(s) in cm}

\item{param}{list of all parameter means (add "_m") and standard deviations (add "_sd") Default parameters are set with very low sd's.
parameters:
\itemize{
\item{Qc_m, Qc_sd:} percentage C of dry mass fish
\item{Qn_m, Qn_sd:} percentage N of dry mass fish
\item{Qp_m, Qp_sd:} percentage P of dry mass fish
\item{Fc_m, Fc_sd:} percentage C of dry mass food
\item{Fn_m, Fn_sd:} percentage N of dry mass food
\item{Fp_m, Fp_sd:} percentage P of dry mass food
\item{AEc_m, AEc_sd:} C-specific assimilation efficiency
\item{AEn_m, AEn_sd:} N-specific assimilation efficiency
\item{AEp_m, AEp_sd:} P-specific assimilation efficiency
\item{Linf_m, Linf_sd:} Von Bertalanffy Growth parameter, theoretical maximum size in TL
\item{k_m, k_sd:} Von Bertalanffy Growth parameter, growth rate
\item{lwa_m, lwa_sd:} Parameter length-weight relationship
\item{lwb_m, lwb_sd:} Parameter length-weight relationship
\item{w_prop_m, wprop_sd:} Ratio between dry weight and wet weight of fish
\item{Tn_m, Tn_sd:} N-specific turnover rate
\item{Tp_m, Tp_sd:} P-specific turnover rate
\item{B0_m, B0_sd:} Normalization constant for resting metabolic rate
\item{a_m, a_sd:} Resting metabolic rate mass-scaling exponent
\item{f_m, f_sd:} Activity scope
\item{asp_m, asp_sd:} Aspect ratio
\item{troph_m, troph_sd:} Trophic level
\item{temp_m, temp_sd:} Temperature
}}

\item{iter}{A positive integer specifying the number of iterations. The default is 2000.}

\item{cor}{A list of correlations between certain parameters: ro_Qc_Qn, ro_Qc_Qp, ro_Qn_Qp,
ro_Fc_Fn, ro_Fc_Fp, ro_Fn_Fp, ro_lwa_lwb, ro_a_B0}

\item{...}{Arguments of rstan::sampling()}
}
\description{
This function combines MTE and stoichiometric theory in order to predict nescessary ingestion and excretion processes. A probability distribution is obtained by including uncertainty of parameters and using MCMC sampling with stan.
}
\details{
Returns a list with two objects: A stanfit object and a dataframe with a summary of all model components.
 See fishflux::extract() to extract a summary of predicted variables and
 fishflux::limitation() to get information on the limiting element.
}
\examples{

model <- fishflux::cnp_model_mcmc(TL = 10, param = list(Qc_m = 40, Qn_m = 10, Qp_m = 4, f_m = 3))
}
\keyword{excretion,}
\keyword{fish,}
\keyword{mcmc}
\keyword{stoichiometry,}
