% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lw.R
\name{find_lw}
\alias{find_lw}
\title{A function to find length-weight relationship parameters a and b}
\usage{
find_lw(sp, mirror = "se")
}
\arguments{
\item{sp}{A charachter value containing the species name}

\item{mirror}{Mirror for fishbase (eg. "se", "org", "us", etc.) Default is}
}
\value{
A dataframe with means and standard deviations of
length-weight parameters
}
\description{
A function to find estimates length-weight relationship parameters available on fishbase. It returns a list of means and standard deviations of a and b obtained from:
\emph{Froese, R., J. Thorson and R.B. Reyes Jr., 2013. A Bayesian approach for estimating length-weight relationships in fishes. J. Appl. Ichthyol. (2013):1-7.}
Please cite Froese et al. (2013), when using these values.
The default mirror for fishbase is set to "se", please change this if needed
}
\examples{
\donttest{
library(fishflux)
library(dplyr)
# find length-weight relationship parameters for one species
find_lw("Lutjanus griseus")

# find length-weight relationship parameters for multiple species and return
#   in data.frame
lapply(c("Chlorurus spilurus","Zebrasoma scopas"), find_lw) |>
  dplyr::bind_rows()
}

}
\keyword{bayesian}
\keyword{fish}
\keyword{fishbase}
\keyword{l-w}
\keyword{relationship}
