% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishflux-package.R
\docType{package}
\name{fishflux-package}
\alias{fishflux-package}
\alias{fishflux}
\title{The 'fishflux' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
