% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishflux.R
\docType{package}
\name{fishflux}
\alias{fishflux}
\alias{fishflux-package}
\title{fishflux: A tool to predict N and P excretion of fish}
\description{
fishflux provides a userfriendly framework to predict
excretion rates of fish. It is based on principles of
metabolic theory and stoichiometry.
}
\details{
It has three main assets:

\itemize{
\item Provides tools to find and check parameters
\item Provides functions to model excretion rates
\item Provides plotting functions to illustrate results}
}
