% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic_parameters.R
\docType{data}
\name{metabolic_parameters}
\alias{metabolic_parameters}
\title{Data with metabolic parameters on family level}
\format{
An object of class \code{data.frame} with 20 rows and 5 columns.
}
\usage{
data(metabolic_parameters)
}
\description{
Data frame containing means and sd for b0 and a for several fish families, extracted from Barneche & Allen (2018)
These paramaters can be used in calculations of metabolic rate in case respirometry data is not available.
}
\examples{
data(metabolic_parameters)
}
\keyword{datasets}
