% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_errors.R
\name{name_errors}
\alias{name_errors}
\title{A function to find errors in fish species names}
\usage{
name_errors(sp)
}
\arguments{
\item{sp}{A vector containing all your scientific species names.}
}
\description{
This function allows you to check if there are errors
in your fish species list and returns inaccurate scientific names
}
\examples{
\donttest{
library(fishflux)
name_errors(c("Chlorurus spilurus", "Zebrasoma scopas"))
name_errors(c("Chlorurus spilurus", "Zebrasoma copas"))}

}
\keyword{fish}
\keyword{names}
\keyword{scientific}
