% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_zebsco.R
\docType{data}
\name{param_zebsco}
\alias{param_zebsco}
\title{List of all parameters needed to run cnp_model for *Zebrasoma scopas*}
\format{
An object of class \code{list} of length 37.
}
\usage{
data(param_zebsco)
}
\description{
List of all parameters needed to run cnp_model for *Zebrasoma scopas*
}
\examples{
data(param_zebsco)
}
\keyword{datasets}
