% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnp.R
\name{plot_cnp}
\alias{plot_cnp}
\title{A function to plot results model}
\usage{
plot_cnp(mod, option = c("overview", "ingestion", "excretion",
  "egestion", "growth", "respiration"), ts = 20, display = c("cnp",
  "c", "n", "p"))
}
\arguments{
\item{mod}{Model output from cnp_model or cnp_model_mcmc.}

\item{option}{Argument to specify which type of model results to plot. Can be either a charachter or a charachter vector with multiple choices.
Choices are one or more of the following:
"ingestion", "excretion", "egestion", "growth", "respiration", or "overview", containing all of the before-mentioned results.
Default is "overview".}

\item{display}{Parameter to specify which element to show result for: "c", "n", "p" or "cnp", which will return three plots.
Default is "cnp".}
}
\description{
This function allows you to plot an overview of the model results in function of the total length of fish
}
\examples{
sp <- "Scarus psittacus"
para <- model_parameters(sp,"Scaridae")$parameters
para
mod1 <- fishflux::cnp_model(TL = 10:19,Fn = 1.3,Fp = 0.1,Fc = 50,
                t0 = para$t0,Linf = para$Linf,k = para$k,
               asp = para$asp,troph = para$troph,f = 3,w_prop = para$w_prop,
               lwa = para$lwa,lwb = para$lwb,temp = 27,N = para$N,P=para$P,C=para$C)
 plot_cnp(test)

mod2 <- cnp_model_mcmc(TL = 5:15, param = list(C_m = 40, N_m = 10, P_m = 4, Fn_sd = 0.05))
plot_cnp(mod2)
}
\keyword{TL}
\keyword{fish,}
\keyword{plot,}
