% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarmi.R
\docType{data}
\name{sarmi}
\alias{sarmi}
\title{Example otolith data *Sargocentron microstoma*}
\format{An object of class \code{data.frame} with 140 rows and 6 columns.}
\usage{
data(sarmi)
}
\description{
Has a otolith data with length and age estimations for 17 individuals.
This data can be used to approximate the VB growth curve.
the data frame contains length, age (from otolith readinga nd back calculations) and id of the fish.
}
\examples{
data(sarmi)
\donttest{fishflux::oto_growth(length = sarmi$length, age = sarmi$age, id = sarmi$id, linf_m = 16, k_m = 0.4)}
}
\keyword{datasets}
