% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_prop.R
\docType{data}
\name{weight_prop}
\alias{weight_prop}
\title{Data frame with dry weight/ wet weight proportions for multiple reef fish families.}
\format{
An object of class \code{data.frame} with 15 rows and 4 columns.
}
\usage{
data(weight_prop)
}
\description{
Data frame with dry weight/ wet weight proportions for multiple reef fish families.
}
\examples{
data(weight_prop)
}
\keyword{datasets}
