% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wprop.R
\name{wprop}
\alias{wprop}
\title{A function to find the ratio of dry weight and wet weight of fish in local database}
\usage{
wprop(family)
}
\arguments{
\item{family}{family}

\item{sp}{Species name}
}
\description{
This function searches the ratio of dry weight and wet weight of fish
on the family level. If the family is not available, an average is returned.
}
\details{
Returns a dataframe with the weight ratio (ww) and it's sd (ww_sd).
}
\examples{

wprop(family="Scaridae")
}
\keyword{cnp_model}
\keyword{find}
\keyword{fish,}
\keyword{for}
\keyword{needed}
\keyword{parameters}
\keyword{some}
