% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zebsco_growth.R
\docType{data}
\name{zebsco_growth}
\alias{zebsco_growth}
\title{Example dataframe for fitting growth curves.}
\format{\describe{
  \item{length}{Total length, in cm}
  \item{age}{age, in years}
  \item{id}{id for each fish}
}}
\usage{
zebsco_growth
}
\description{
length-age data from otolith readings and back-calculation for a coral reef fish, Zebrasoma scopas
}
\keyword{datasets}
