% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnp_model_mcmc.R
\name{cnp_model_mcmc}
\alias{cnp_model_mcmc}
\title{A function to predict N and P excretion, CNP egestion, CNP ingestion rate, using MCMC and stan}
\usage{
cnp_model_mcmc(TL, param, iter = 1000, ...)
}
\arguments{
\item{TL}{total length of a fish in cm}

\item{param}{list of all parameter means ("_m") and standard deviations ("_sd") Default parameters are set with very low sd's. See \link[fishflux]{cnp_model}  for a list of all requested parameters}

\item{iter}{A positive integer specifying the number of iterations. The default is 2000.}

\item{...}{Arguments of rstan::sampling()}
}
\description{
This function combines MTE and stoichiometric theory in order to predict nescessary ingestion and excretion processes. A probability distribution is obtained by including uncertainty of parameters and using MCMC sampling with stan.
}
\details{
Returns a list with three objects: A stanfit object, a dataframe with a summary of all model components and a vector containing the limiting element.
}
\examples{

model <- fishflux::cnp_model_mcmc(TL = 10:12, param = list(C_m = 40, N_m = 10, P_m = 4, f_m = 3))
}
\keyword{excretion,}
\keyword{fish,}
\keyword{mcmc}
\keyword{stoichiometry,}
