% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{A function to extract specific model output parameters from result}
\usage{
extract(mod, par)
}
\arguments{
\item{mod}{Output from cnp_mod_mcmc()}

\item{par}{Character vector specifying which output parameter that should be returned}
}
\description{
A function to extract specific model output parameters from result
}
\details{
Returns a dataframe with a summary of the selected output parameters
}
\examples{

model <- fishflux::cnp_model_mcmc(TL = 5:10, param = list(C_m = 40, N_m = 10, P_m = 4))
fishflux::extract(model, c("N_ex","P_ex"))


}
\keyword{excretion,}
\keyword{fish,}
\keyword{mcmc}
\keyword{stoichiometry,}
