% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishflux-package.R
\docType{package}
\name{fishflux-package}
\alias{fishflux-package}
\alias{fishflux}
\title{The 'fishflux' package.}
\description{
The \code{fishflux} package provides a tool to model fluxes of C (carbon),
N (nitrogen) and P (phosphorus) in fish. It combines basic principles from
elemental stoichiometry and metabolic theory. The package offers a
user-friendly interface to make nutrient dynamic modelling available for
anyone. \code{fishflux} is mostly targeted towards fish ecologists, wishing
to predict nutrient ingestion, egestion and excretion to study fluxes of
nutrients and energy.
Main assets:
\itemize{
\item Provides functions to model fluxes of Carbon, Nitrogen and Phosphorus for fish with or without the MCMC sampler provided by stan.
\item Provides some tools to find the right parameters as inputs into the model
\item Provides a plotting function to illustrate results
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nschiett.github.io/fishflux/}
  \item Report bugs at \url{https://github.com/nschiett/fishflux/issues/}
}

}
\author{
\strong{Maintainer}: Nina M. D. Schiettekatte \email{nina.schiettekatte@gmail.com}

Authors:
\itemize{
  \item Diego Barneche \email{d.barneche@aims.gov.au}
}

}
