% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishflux.R
\docType{package}
\name{fishflux}
\alias{fishflux}
\alias{fishflux-package}
\title{fishflux: A tool to model nutrient fluxes in fish}
\description{
The `r fishflux` package provides a tool to model fluxes of C (carbon), N (nitrogen) and P (phosphorus) in fish.
It combines basic priciples from elemental stoichiometry and metabolic theory.
The package offers a userfriendly interface to make nutrient dynamic modelling available for anyone.
`r fishflux` is mostly targeted towards fish ecologists,
wishing to predict nutrient ingestion, egestion and excretion to study fluxes of nutrients and energy.
Main assets:
\itemize{
\item Provides functions to model fluxes of Carbon, Nitrogen and Phosphorus for fish with or without the MCMC sampler provided by stan.
\item Provides some tools to find the right parameters as inputs into the model
\item Provides a plotting function to illustrate results
}
}
