% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oto_growth.R
\name{oto_growth}
\alias{oto_growth}
\title{A function to extract growth parameters from otolith data}
\usage{
oto_growth(length, age, id, lmax = 20, linf_m, plot = TRUE, ...)
}
\arguments{
\item{length}{Numerical vector with length}

\item{age}{Numerical vector with age}

\item{id}{Character vector with fish id}

\item{lmax}{maximum size. Based on this value, maximum growth rate kmax will be computed.}

\item{linf_m}{Prior for linf}

\item{plot}{option to plot model fit (TRUE or FALSE)}

\item{...}{Additional arguments, see ?rstan::sampling()}
}
\description{
A function to extract growth parameters from otolith data
}
\details{
Returns a dataframe with estimates for linf, k and t0, sl and gp.
There is a hierarchical structure for linf and k, so that there is a unique estimate for these parameters per individual (linf_j, k_j).
linf and k are the population level estimates of linf and k. kmax is the standardised growth parameter, depending on lmax
(kmax = exp(sl * log(lmax) + gp), see Morais and Bellwood (2018) for details)
}
\examples{

zs <- fishflux::zebsco_growth
fishflux::oto_growth(length = zs$length, age = zs$age, id = zs$id, lmax = 25, linf_m = 15, iter = 2000, chains = 1)

}
\keyword{Bertalanfy}
\keyword{Von}
\keyword{fish,}
\keyword{growth,}
