% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oto_growth.R
\name{oto_growth}
\alias{oto_growth}
\title{A function to extract growth parameters from otolith data}
\usage{
oto_growth(length, age, id, linf_min = 0, linf_m, k_m = 0.5,
  iter = 2000, warmup = 1000)
}
\arguments{
\item{length}{Numerical vector with length}

\item{age}{Numerical vector with age}

\item{id}{Character vector with fish id}

\item{linf_min}{possibility to add a minimum for linf, default is 0}

\item{linf_m}{Prior for linf}

\item{k_m}{Prior for k, default is 0.5}
}
\description{
A function to extract growth parameters from otolith data
}
\details{
Returns a dataframe with estimates for Linf, k and t0
}
\examples{

sarmi <- fishflux::sarmi
oto_growth(length = sarmi$length, age = sarmi$age, id = sarmi$id, linf_m = 16, k_m = 0.4, iter = 100, warmup = 10)

}
\keyword{Bertalanfy}
\keyword{Von}
\keyword{fish,}
\keyword{growth,}
